app.js
This is a basic program that allows people to create user profiles and
interact with other profiles. Users can attack other players and/or
give them points. There is a hierarchy of user type that dictates the
outcomes of certain actions. ‘Super’ users cannot attack ‘superDuper’
users and if a ‘standard’ user attacks a ‘super’ or ‘superDuper’ user,
the attacker loses half of their points as opposed to the traditional
attack function where the attacker is awarded. JavaScript concepts
touched in this program include, functions, objects, prototyping and
conditional statements among others.